Conjunctions & Disjunctions should return all matching intervals
The ConjunctionIntervalIterator and DisjunctionIntervalIterator return spanning
intervals for their parent queries, but in those cases we don't actually want
spans, we want the individual matches for each clause.  This commit introduces a
CombinedIntervalIterator that does just that, and updates the various
Conjunction and Disjunction scorers to use it.  UnorderedNearQuery explicitly
creates a CII for it's filter.